Icing stab

+++ src/math/bounds.rs
@@ -33,7 +33,7 @@ impl Bounds {
         res
     }
     // Вспомогательный конструктор 
-    pub fn from_frames(frames: &Vec<(usize, f64)>) -> Self {
+    pub fn from_frames(frames: &Vec<(i32, f64)>) -> Self {
         assert!(frames.len() > 1, "frames.len() {:?} > 1", frames);
         let mut res = Vec::new();
         let shift = if frames[0].1 >= 0. {
